+++ src/privacy-keys/index.ts
@@ -1,7 +1,6 @@
 export {
     deriveSpendingKeyTypedData,
     deriveSpendingKeyMessageV2,
-    deriveSpendingKeyMessage,
     SPENDING_KEY_VERIFYING_CONTRACT,
     SPENDING_KEY_WARNING,
 } from './SpendingKeyRequest';
@@ -14,5 +13,4 @@ export {
     deriveOwnerPk,
     MIN_SIGNATURE_BYTES,
 } from './PrivacyKeys';
-export type { KeyVersion } from './PrivacyKeys';
 export { PrivacyKeyManager } from './PrivacyKeyManager';